requirements.txt
flask
flask_sqlalchemy
flask_bcrypt
flask_mail
flask_login
flask_wtf
wtforms
itsdangerous
Pillow
psprint
email-validator
pyyaml
python-dateutil
